Refactor migrate ng v9 to v10 (#53)

+++ decorate-angular-cli.js
@@ -1,5 +1,5 @@
 /**
- * This file decorate the Angular CLI with the Nx CLI to enable features such as computation caching,
+ * This file decorates the Angular CLI with the Nx CLI to enable features such as computation caching
  * and faster execution of tasks.
  *
  * It does this by:
@@ -11,7 +11,7 @@
  * The Nx CLI decorates the Angular CLI, so the Nx CLI is fully compatible with it.
  * Every command you run should work the same when using the Nx CLI, except faster.
  *
- * Because of symliking you can still type `ng build/test/lint` in the terminal. The ng command, in this case,
+ * Because of symlinking you can still type `ng build/test/lint` in the terminal. The ng command, in this case,
  * will point to nx, which will perform optimizations before invoking ng. So the Angular CLI is always invoked.
  * The Nx CLI simply does some optimizations before invoking the Angular CLI.
  *
@@ -66,7 +66,8 @@ function symlinkNgCLItoNxCLI() {
        * Such that it works in all shells and works with npx.
        */
       ['', '.cmd', '.ps1'].forEach(ext => {
-        fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
+        if (fs.existsSync(nxPath + ext))
+          fs.writeFileSync(ngPath + ext, fs.readFileSync(nxPath + ext));
       });
     } else {
       // If unix-based, symlink
@@ -83,7 +84,11 @@ function symlinkNgCLItoNxCLI() {
 try {
   symlinkNgCLItoNxCLI();
   patchAngularCLI(angularCLIInitPath);
-  output.log({ title: 'Angular CLI has been decorated to enable computation caching.' });
+  output.log({
+    title: 'Angular CLI has been decorated to enable computation caching.',
+  });
 } catch (e) {
-  output.error({ title: 'Decoration of the Angular CLI did not complete successfully' });
+  output.error({
+    title: 'Decoration of the Angular CLI did not complete successfully',
+  });
 }
